api/init_db.py
import psycopg2
import os
from dotenv import load_dotenv

# Carregar variáveis do arquivo .env
load_dotenv()

print("Conectando no banco de dados...")

# Conectar no banco
conn = psycopg2.connect(
    host=os.getenv('DB_HOST'),
    port=os.getenv('DB_PORT'),
    database=os.getenv('DB_NAME'),
    user=os.getenv('DB_USER'),
    password=os.getenv('DB_PASSWORD')
)

cursor = conn.cursor()

print("Criando tabela 'comments'...")

# Criar tabela comments
cursor.execute("""
    CREATE TABLE IF NOT EXISTS comments (
        id SERIAL PRIMARY KEY,
        email VARCHAR(255) NOT NULL,
        comment TEXT NOT NULL,
        content_id VARCHAR(100) NOT NULL,
        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
    );
""")

conn.commit()
cursor.close()
conn.close()

print("✅ Tabela 'comments' criada com sucesso!")
