api/requirements.txt
flask==3.0.0
prometheus-flask-exporter==0.23.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
